lexora-ai/pyproject.toml
[project]
name = "lexora-ai"
version = "0.1.0"
description = "Lexora AI - Legal Copilot for Uzbek Law"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.4.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.6.0",
    "pytest>=8.0.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]